backend/requirements/base.txt
asgiref==3.6.0
Django==4.1.7
Pillow==9.4.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
config==0.5.1
env_utils==1.2.0
netcat==0.0.1a0
django-staticfiles==1.2.1
drf-nested-routers==0.93.4
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2
httpx==0.24.1
model_bakery==1.11.0
Faker==18.9.0
celery==5.2.7
redis==4.5.5
django-solo==2.1.0
urllib3==1.26.7
mysqlclient==2.2.0